Added Routes and Views
