+++ Server/app/templates/index.html
@@ -1,6 +1,7 @@
 <html>
 <head>
     <title>Login Screen</title>
+    <meta http-equiv=“refresh" content="5">
 </head>
 <body>
     <h1>Login Screen</h1>
